% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{getCspUserName}
\alias{getCspUserName}
\title{获取导购员名称}
\usage{
getCspUserName(conn = tsda::conn_rds("rdbe"), app_id = "caas")
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序}
}
\value{
返回列表
}
\description{
获取导购员名称
}
\examples{
getCspUserName()
}
