% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userManagementRpt.R
\name{getNewUsers}
\alias{getNewUsers}
\title{判断新用户}
\usage{
getNewUsers(conn = conn_rds("rdbe"), app_id = "caas", users)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序}

\item{users}{用户}
}
\value{
返回逻辑值
}
\description{
判断新用户
}
\examples{
getNewUsers()
}
