% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qianNiu.R
\name{getQianNiuRpt}
\alias{getQianNiuRpt}
\title{查询客服报表}
\usage{
getQianNiuRpt(
  conn = conn_rds_nsic(),
  FStartDate = "2020-04-28",
  FEndDate = "2020-05-07",
  FUser = "demo"
)
}
\arguments{
\item{conn}{连接}

\item{FStartDate}{开始日期}

\item{FEndDate}{结束日期}

\item{FUser}{导购员}
}
\value{
返回数据
}
\description{
查询客服报表
}
\examples{
getCspRpt()
}
