% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qianNiu.R
\name{getQN_logCount_byCspName}
\alias{getQN_logCount_byCspName}
\title{返回日志指定日期记录条数}
\usage{
getQN_logCount_byCspName(
  conn = conn_rds_nsic(),
  FUser = "admin",
  FDate = "2020-05-01"
)
}
\arguments{
\item{conn}{连接}

\item{FUser}{用户}

\item{FDate}{日期}
}
\value{
返回值
}
\description{
返回日志指定日期记录条数
}
\examples{
getQN_logCount_byCspName()
}
