% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tspMngr.R
\name{getTspRpt}
\alias{getTspRpt}
\title{查询TSP结果}
\usage{
getTspRpt(
  conn = conn_rds_nsic(),
  FStartDate = "2020-04-28",
  FEndDate = "2020-05-07"
)
}
\arguments{
\item{conn}{连接}

\item{FStartDate}{开始}

\item{FEndDate}{结束}
}
\value{
返回值
}
\description{
查询TSP结果
}
\examples{
getTspRpt()
}
