% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userManagementRpt.R
\name{getUserInfoRpt}
\alias{getUserInfoRpt}
\title{查询用户信息}
\usage{
getUserInfoRpt(conn = conn_rds("rdbe"), app_id = "caas")
}
\arguments{
\item{conn}{连接信息}

\item{app_id}{程序名称}
}
\value{
返回值
}
\description{
查询用户信息
}
\examples{
getUserInfoRpt
}
