% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qianNiu.R
\name{log_compressIntoQA}
\alias{log_compressIntoQA}
\title{针对日志处理QA对}
\usage{
log_compressIntoQA(
  conn = conn_rds_nsic(),
  FUser = "腊梅",
  log_date = "2020-08-19"
)
}
\arguments{
\item{conn}{连接}

\item{FUser}{用户}

\item{log_date}{日志}
}
\value{
返回值
}
\description{
针对日志处理QA对
}
\examples{
log_compressIntoQA()
}
