% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qianNiu.R
\name{log_getUserListByDate}
\alias{log_getUserListByDate}
\title{根据日期获取用户}
\usage{
log_getUserListByDate(conn = conn_rds_nsic(), log_date = "2020-08-19")
}
\arguments{
\item{conn}{连接}

\item{log_date}{日志日期}
}
\value{
返回值
}
\description{
根据日期获取用户
}
\examples{
log_getUserListByDate()
}
