% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qianNiu.R
\name{log_qaPair_byDate}
\alias{log_qaPair_byDate}
\title{按日期处理QA表}
\usage{
log_qaPair_byDate(conn = conn_rds_nsic(), log_date = "2020-08-19")
}
\arguments{
\item{conn}{连接}

\item{log_date}{日期}
}
\value{
返回值
}
\description{
按日期处理QA表
}
\examples{
log_qaPair_byDate()
}
