% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qianNiu.R
\name{log_qaPair_reboot}
\alias{log_qaPair_reboot}
\title{重新初始化相关数据}
\usage{
log_qaPair_reboot(conn = conn_rds_nsic())
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
重新初始化相关数据
}
\examples{
log_qaPair_reboot()
}
