% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qna.R
\name{log_qna_combined}
\alias{log_qna_combined}
\title{针对QA对进行合并}
\usage{
log_qna_combined(
  conn = conn_rds_nsic(),
  FUser = "蓝淀",
  log_date = "2020-08-18",
  sep = "|"
)
}
\arguments{
\item{conn}{连接}

\item{FUser}{用户}

\item{log_date}{日期}

\item{sep}{定义分隔符}
}
\value{
返回日
}
\description{
针对QA对进行合并
}
\examples{
log_qna_combined()
}
