% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qna.R
\name{log_qna_combined3}
\alias{log_qna_combined3}
\title{针对QA对进行合并3}
\usage{
log_qna_combined3(
  conn = conn_rds_nsic(),
  FUser = "腊梅",
  log_date = "2020-08-19",
  sep = "|"
)
}
\arguments{
\item{conn}{连接}

\item{FUser}{用户}

\item{log_date}{日期}

\item{sep}{定义分隔符}
}
\value{
返回日
}
\description{
针对QA对进行合并3
}
\examples{
log_qna_combined3()
}
