% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qna.R
\name{log_qna_combined_byDate3}
\alias{log_qna_combined_byDate3}
\title{按天对数据进行合并3}
\usage{
log_qna_combined_byDate3(
  conn = conn_rds_nsic(),
  log_date = "2020-08-19",
  sep = "|"
)
}
\arguments{
\item{conn}{连接}

\item{log_date}{日期}

\item{sep}{分隔符}
}
\value{
返回值
}
\description{
按天对数据进行合并3
}
\examples{
log_qna_combined_byDate3()
}
