% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qna.R
\name{log_qna_updatedLogTag}
\alias{log_qna_updatedLogTag}
\title{针对日志数据进行打标更新}
\usage{
log_qna_updatedLogTag(conn = conn_rds_nsic())
}
\arguments{
\item{conn}{链接信息}
}
\value{
返回值
}
\description{
针对日志数据进行打标更新
}
\examples{
log_qna_updatedLogTag()
}
