% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qna.R
\name{log_qna_updatedLogTag3}
\alias{log_qna_updatedLogTag3}
\title{针对日志数据进行打标更新3}
\usage{
log_qna_updatedLogTag3(conn = conn_rds_nsic())
}
\arguments{
\item{conn}{链接信息}
}
\value{
返回值
}
\description{
针对日志数据进行打标更新3
}
\examples{
log_qna_updatedLogTag3()
}
