% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oilCard.R
\name{oilCard_backup_del}
\alias{oilCard_backup_del}
\title{删除库中数据}
\usage{
oilCard_backup_del(conn = tsda::conn_rds("nsic"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
删除库中数据
}
\examples{
oilCard_backup_del()
}
