% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oilCard.R
\name{oilCard_formatDeliverDate}
\alias{oilCard_formatDeliverDate}
\title{处理油卡发货日期异常数据}
\usage{
oilCard_formatDeliverDate(x)
}
\arguments{
\item{x}{针对日期进行处理}
}
\value{
返回值
}
\description{
处理油卡发货日期异常数据
}
\examples{
oilCard_formatDeliverDate()
}
