% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oilCard.R
\name{oilCard_formatDeliverDates}
\alias{oilCard_formatDeliverDates}
\title{针对日期数据进行批量处理}
\usage{
oilCard_formatDeliverDates(data)
}
\arguments{
\item{data}{数据}
}
\value{
返回值
}
\description{
针对日期数据进行批量处理
}
\examples{
oilCard_formatDeliverDates()
}
