% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oilCard.R
\name{oilCard_readExcel}
\alias{oilCard_readExcel}
\title{读取油卡数据}
\usage{
oilCard_readExcel(file = "data-raw/oilCardData.xlsx")
}
\arguments{
\item{file}{文件名}
}
\value{
返回值
}
\description{
读取油卡数据
}
\examples{
oilCard_readExcel()
}
