% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oilCard.R
\name{oilCard_writeDB}
\alias{oilCard_writeDB}
\title{油卡数据写入数据库}
\usage{
oilCard_writeDB(
  file = "data-raw/oilCardData.xlsx",
  conn = tsda::conn_rds("nsic")
)
}
\arguments{
\item{file}{文件}

\item{conn}{连接}
}
\value{
返回值
}
\description{
油卡数据写入数据库
}
\examples{
oilCard_writeDB()
}
