% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oilCard.R
\name{oildCard_selectDB}
\alias{oildCard_selectDB}
\title{查询油卡}
\usage{
oildCard_selectDB(conn = tsda::conn_rds("nsic"), FKeyWord = "ljiang1469")
}
\arguments{
\item{conn}{连接}

\item{FKeyWord}{关键词}
}
\value{
返回值
}
\description{
查询油卡
}
\examples{
oildCard_selectDB()
}
