% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oilCard.R
\name{oildCard_selectDB2}
\alias{oildCard_selectDB2}
\title{增值油卡订单查询功能}
\usage{
oildCard_selectDB2(conn = tsda::conn_rds("nsic"), FKeyWord = "ljiang1469")
}
\arguments{
\item{conn}{连接}

\item{FKeyWord}{关键词}
}
\value{
返回值
}
\description{
增值油卡订单查询功能
}
\examples{
oildCard_selectDB2()
}
