% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oilCard.R
\name{oildCard_selectDB_all}
\alias{oildCard_selectDB_all}
\title{查询油卡数据}
\usage{
oildCard_selectDB_all(conn = tsda::conn_rds("nsic"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
查询油卡数据
}
\examples{
oildCard_selectDB_all()
}
