% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{page_setting}
\alias{page_setting}
\title{针对写入数据进行分页处理}
\usage{
page_setting(totalRow = 14, pageRow = 4)
}
\arguments{
\item{totalRow}{总行数}

\item{pageRow}{每页数}
}
\value{
返回值
}
\description{
针对写入数据进行分页处理
}
\examples{
page_setting()
}
