% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qna.R
\name{qna_getUserListByDate3}
\alias{qna_getUserListByDate3}
\title{按天获取相关的用户3}
\usage{
qna_getUserListByDate3(conn = conn_rds_nsic(), log_date = "2020-08-19")
}
\arguments{
\item{conn}{连接}

\item{log_date}{日志日期}
}
\value{
返回值
}
\description{
按天获取相关的用户3
}
\examples{
qna_getUserListByDate3()
}
