% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnlogBlackList.R
\name{qnlog_blackList_read}
\alias{qnlog_blackList_read}
\title{读取日志黑名单}
\usage{
qnlog_blackList_read(file = "data-raw/顾客问题_黑名单_reviewed.xlsx")
}
\arguments{
\item{file}{日志黑名单文件}
}
\value{
返回值
}
\description{
读取日志黑名单
}
\examples{
qnlog_blackList_read()
}
