% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnlogBlackList.R
\name{qnlog_blackList_writeDB}
\alias{qnlog_blackList_writeDB}
\title{上传千牛日志黑名单}
\usage{
qnlog_blackList_writeDB(
  conn = tsda::conn_rds("nsic"),
  file = "data-raw/顾客问题_黑名单_reviewed.xlsx"
)
}
\arguments{
\item{conn}{连接}

\item{file}{文件}
}
\value{
返回值
}
\description{
上传千牛日志黑名单
}
\examples{
qnlog_blackList_writeDB()
}
