% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnlogBlackList.R
\name{qnlog_blackListEq_read}
\alias{qnlog_blackListEq_read}
\title{读取精确匹配黑名单}
\usage{
qnlog_blackListEq_read(file = "data-raw/顾客问题_黑名单_reviewed.xlsx")
}
\arguments{
\item{file}{文件名}
}
\value{
返回值
}
\description{
读取精确匹配黑名单
}
\examples{
qnlog_blackListEq_read()
}
