% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnlogBlackList.R
\name{qnlog_blackListEq_writeDB}
\alias{qnlog_blackListEq_writeDB}
\title{写入严格匹配黑名单到数据库}
\usage{
qnlog_blackListEq_writeDB(
  conn = tsda::conn_rds("nsic"),
  file = "data-raw/顾客问题_黑名单_reviewed.xlsx"
)
}
\arguments{
\item{conn}{连接}

\item{file}{文件}
}
\value{
返回值
}
\description{
写入严格匹配黑名单到数据库
}
\examples{
qnlog_blackListEq_writeDB()
}
