% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnlogBlackList.R
\name{qnlog_getBlackList_eq}
\alias{qnlog_getBlackList_eq}
\title{获取最新的黑名单}
\usage{
qnlog_getBlackList_eq(conn = tsda::conn_rds("nsic"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取最新的黑名单
}
\examples{
qnlog_getBlackList_eq()
}
