% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnlogBlackList.R
\name{qnlog_getLog_perUserDate}
\alias{qnlog_getLog_perUserDate}
\title{按用户按天获取日志问信息}
\usage{
qnlog_getLog_perUserDate(
  conn = tsda::conn_rds("nsic"),
  FUser = "腊梅",
  log_date = "2020-05-03"
)
}
\arguments{
\item{conn}{连接}

\item{FUser}{导购}

\item{log_date}{日期}
}
\value{
返回值
}
\description{
按用户按天获取日志问信息
}
\examples{
qnlog_getLog_perUserDate()
}
