% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnlogBlackList.R
\name{qnlog_getUserDate}
\alias{qnlog_getUserDate}
\title{获取有效问的作者与日期数据}
\usage{
qnlog_getUserDate(
  conn = tsda::conn_rds("nsic"),
  FStartDate = "2020-06-06",
  FEndDate = "2020-06-14"
)
}
\arguments{
\item{conn}{连接}

\item{FStartDate}{开始日期}

\item{FEndDate}{结束日期}
}
\value{
返回值
}
\description{
获取有效问的作者与日期数据
}
\examples{
qnlog_getUserDate()
}
