% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnlogBlackList.R
\name{qnlog_logTag_delete3}
\alias{qnlog_logTag_delete3}
\title{删除原有的日志信息3}
\usage{
qnlog_logTag_delete3(
  conn = tsda::conn_rds("nsic"),
  FUser = "腊梅",
  log_date = "2020-05-03"
)
}
\arguments{
\item{conn}{连接}

\item{FUser}{用户}

\item{log_date}{日志}
}
\value{
无返回值
}
\description{
删除原有的日志信息3
}
\examples{
qnlog_logTag_delete()
}
