% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnlogBlackList.R
\name{qnlog_logTag_read3}
\alias{qnlog_logTag_read3}
\title{获取单个数据进行打款3}
\usage{
qnlog_logTag_read3(
  conn = tsda::conn_rds("nsic"),
  FUser = "腊梅",
  log_date = "2020-05-03",
  dict_contain = qnlog_getBlackList(),
  dict_eq = qnlog_getBlackList_eq()
)
}
\arguments{
\item{conn}{连接}

\item{FUser}{用户}

\item{log_date}{日期}

\item{dict}{字词}
}
\value{
返回值
}
\description{
获取单个数据进行打款3
}
\examples{
qnlog_logTag_read3()
}
