% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnlogBlackList.R
\name{qnlog_logTag_readDB}
\alias{qnlog_logTag_readDB}
\title{读取打标日志内容}
\usage{
qnlog_logTag_readDB(
  conn = tsda::conn_rds("nsic"),
  FStartDate = "2020-06-08",
  FEndDate = "2020-06-14"
)
}
\arguments{
\item{conn}{连接}

\item{FStartDate}{开始日期}

\item{FEndDate}{结束日期}
}
\value{
返回值
}
\description{
读取打标日志内容
}
\examples{
qnlog_logTag_readDB()
}
