% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnlogBlackList.R
\name{qnlog_logTag_writeDB3}
\alias{qnlog_logTag_writeDB3}
\title{日志打标信息写入数据库}
\usage{
qnlog_logTag_writeDB3(
  conn = tsda::conn_rds("nsic"),
  FUser = "腊梅",
  log_date = "2020-05-03",
  dict_contain = qnlog_getBlackList(),
  dict_eq = qnlog_getBlackList_eq()
)
}
\arguments{
\item{conn}{连接}

\item{FUser}{用户}

\item{log_date}{日志}

\item{dict_contain}{模糊匹配}

\item{dict_eq}{严格匹配}
}
\value{
返回值
}
\description{
日志打标信息写入数据库
}
\examples{
qnlog_logTag_writeDB3()
}
