% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnlogBlackList.R
\name{qnlog_logTagBatch_writeDB}
\alias{qnlog_logTagBatch_writeDB}
\title{针对日志打标进行处理批量处理}
\usage{
qnlog_logTagBatch_writeDB(
  conn = tsda::conn_rds("nsic"),
  show_process = FALSE,
  FStartDate = "2020-06-08",
  FEndDate = "2020-06-14",
  dict_contain = qnlog_getBlackList(),
  dict_eq = qnlog_getBlackList_eq()
)
}
\arguments{
\item{conn}{连接}

\item{show_process}{是否显示进度条}

\item{FStartDate}{开始日期}

\item{FEndDate}{结束日期}

\item{dict_contain}{模糊匹配}

\item{dict_eq}{精确匹配}
}
\value{
返回值
}
\description{
针对日志打标进行处理批量处理
}
\examples{
qnlog_logTagBatch_writeDB()
}
