% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qianNiu.R
\name{getQN_log_byCspName}
\alias{getQN_log_byCspName}
\title{查询千牛日志按导购人员}
\usage{
getQN_log_byCspName(conn = conn_rds_nsic(), FUser = "admin")
}
\arguments{
\item{conn}{连接}

\item{FUser}{用户}
}
\value{
返回数据框
}
\description{
查询千牛日志按导购人员
}
\examples{
getQN_log_byCspName()
}
